开始公会Controller

+++ gameserver/src/main/java/com/wan37/gameServer/game/trade/controller/TradeGoodsController.java
@@ -31,7 +31,7 @@ public class TradeGoodsController implements IController {
     @Override
     public void handle(ChannelHandlerContext ctx, Message message) {
         String[] args =  new String(message.getContent()).split("\\s+");
-        Long itemId = Long.valueOf(args[1]);
+        Integer itemId = Integer.valueOf(args[1]);
 
         tradeService.tradeGoods(ctx,itemId);
     }
